requirements.txt
# Document extraction
PyMuPDF>=1.23.0
python-pptx>=0.6.0

# Chunking / token counting
tiktoken>=0.5.0

# Progress display
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.0.0

# Vector store
chromadb>=1.5.5

# Knowledge graph
kuzu>=0.11.3
rapidfuzz>=3.14.0

# LM Studio API client
openai>=1.0.0

# Multi-provider LLM and embedding adapter
litellm>=1.45.0
python-dotenv>=1.0.0

# Chat UI (archived — replaced by adk web)
streamlit>=1.30.0

# Google ADK multi-agent framework
google-adk>=1.0.0

# BM25 keyword index for hybrid retrieval (RAG-01)
rank_bm25>=0.2.2

# Cross-encoder reranker (RAG-02)
sentence-transformers>=2.7.0
